Extract `fileDownload` in `FileUtil` to download anywhere
