+++ app/src/main/java/kr/khs/oneboard/utils/FileUtil.kt
@@ -1,10 +1,14 @@
 package kr.khs.oneboard.utils
 
 import android.annotation.SuppressLint
+import android.app.DownloadManager
 import android.content.ContentResolver
+import android.content.Context
 import android.database.Cursor
 import android.net.Uri
+import android.os.Environment
 import android.provider.OpenableColumns
+import androidx.fragment.app.Fragment
 import okhttp3.MediaType
 import okhttp3.MediaType.Companion.toMediaType
 import okhttp3.MultipartBody
@@ -55,4 +59,27 @@ fun Uri.getFileName(contentResolver: ContentResolver): String {
         result = this.lastPathSegment
     }
     return result ?: "nullll"
+}
+
+fun Fragment.fileDownload(
+    title: String,
+    description: String,
+    downloadUrl: String,
+    fileName: String
+) {
+    val request = DownloadManager.Request(Uri.parse(downloadUrl))
+        .setTitle(title)
+        .setDescription(description)
+        .setNotificationVisibility(DownloadManager.Request.VISIBILITY_VISIBLE_NOTIFY_COMPLETED)
+        .setDestinationInExternalPublicDir(
+            Environment.DIRECTORY_DOWNLOADS,
+            fileName
+        )
+        .setRequiresCharging(false)
+        .setAllowedOverMetered(true)
+        .setAllowedOverRoaming(true)
+
+    (requireContext().getSystemService(Context.DOWNLOAD_SERVICE) as DownloadManager).enqueue(
+        request
+    )
 }
